remove extra wrap commonent
