+++ web/package.json
@@ -15,14 +15,10 @@
     "@svgr/webpack": "^5.5.0",
     "apollo-cache-inmemory": "^1.6.6",
     "apollo-client": "^2.6.10",
-    "apollo-link": "^1.2.14",
     "apollo-link-error": "^1.1.13",
-    "apollo-link-http": "^1.5.17",
     "dotenv": "^8.2.0",
     "graphql": "^15.4.0",
-    "isomorphic-unfetch": "^3.1.0",
     "next": "10.0.4",
-    "next-with-apollo": "^5.1.0",
     "react": "17.0.1",
     "react-device-detect": "^1.15.0",
     "react-dom": "17.0.1"
